requirements.txt
chromadb
sentence-transformers
umap-learn
plotly
pandas
scikit-learn
pytest
langchain-text-splitters
pypdf
google-genai
flask
python-dotenv
google-adk
ragas 
langchain-google-genai
rapidfuzz